refactor: move elements as dialogs instead of expansion panels
