+++ component/widget/export_dialog.py
@@ -3,9 +3,7 @@ from pathlib import Path
 
 import ee
 import rasterio as rio
-from ipyleaflet import WidgetControl
 from matplotlib.colors import to_rgba
-from sepal_ui import mapping as sm
 from sepal_ui import sepalwidgets as sw
 from sepal_ui.scripts import gee
 from sepal_ui.scripts import utils as su
@@ -17,8 +15,13 @@ from component.message import cm
 ee.Initialize()
 
 
-class ExportMap(WidgetControl):
+class ExportMapDialog(sw.Dialog):
     def __init__(self, **kwargs):
+        self.v_model = False
+        self.max_width = "700px"
+        # init the dialog
+        super().__init__(**kwargs)
+
         # init the downloadable informations
         self.geometry = None
         self.dataset = None
@@ -40,9 +43,9 @@ class ExportMap(WidgetControl):
         # add alert and btn component for the loading_button
         self.alert = sw.Alert()
         self.btn = sw.Btn(cm.export.apply, small=True)
-        title = sw.Html(tag="h3", children=[cm.export.title], class_="pl-1")
-        close_btn = sw.Icon(children=["fas fa-times"], small=True)
-        sw.CardTitle(children=[title, sw.Spacer(), close_btn])
+
+        title = sw.CardTitle(children=[cm.export.title])
+
         text = sw.CardText(
             children=[
                 w_scale_lbl,
@@ -52,31 +55,18 @@ class ExportMap(WidgetControl):
                 self.alert,
             ]
         )
-        action = sw.CardActions(children=[self.btn])
-        export_data = sw.Card(children=[title, text, action], class_="pt-1")
-        self.download_btn = sm.MapBtn("mdi-cloud-download", v_on="menu.on")
-        slot = {"name": "activator", "variable": "menu", "children": self.download_btn}
+        actions = sw.CardActions(children=[sw.Spacer(), self.btn])
 
-        # create the menu
-        menu = sw.Menu(
-            v_model=False,
-            value=False,
-            close_on_content_click=False,
-            # nudge_width=200,
-            offset_x=True,
-            children=[export_data],
-            v_slots=[slot],
-            top="bottom" in kwargs["position"],
-            bottom="top" in kwargs["position"],
-            left="right" in kwargs["position"],
-            right="left" in kwargs["position"],
-        )
+        self.children = [sw.Card(children=[title, text, actions])]
 
-        super().__init__(widget=menu, **kwargs)
+        super().__init__()
 
         # add js behaviour
         self.btn.on_event("click", self._apply)
-        close_btn.on_event("click", lambda *_: setattr(self.menu, "v_model", False))
+
+    def open_dialog(self, *_):
+        """Open dialog."""
+        self.v_model = True
 
     def set_data(self, dataset, geometry, name, aoi_name):
         """set the dataset and the geometry to allow the download."""
